proposed removal of serializable attribute(s) where no longer necessary
