+++ src/Common/src/NativeMethods.cs
@@ -2458,7 +2458,6 @@ namespace System.Windows.Forms
         }
 
         [StructLayout(LayoutKind.Sequential)]
-        [Serializable]
         public struct MSG
         {
             public IntPtr hwnd;
@@ -3315,8 +3314,8 @@ namespace System.Windows.Forms
             public int[] widths = new int[6];
         }
 
+        // TODO: determine if class or struct
         [StructLayout(LayoutKind.Sequential)]
-        [Serializable]
         public class MSOCRINFOSTRUCT
         {
             public int cbSize = Marshal.SizeOf<MSOCRINFOSTRUCT>();              // size of MSOCRINFO structure in bytes.
