+++ src/System.Windows.Forms/src/System/Windows/Forms/ListViewItem.cs
@@ -15,6 +15,8 @@ using System.Runtime.Serialization;
 
 namespace System.Windows.Forms
 {
+// TODO: add serializable reasoning (@Devendar)
+
     /// <summary>
     /// Implements an item of a <see cref='System.Windows.Forms.ListView'/>.
     /// </summary>
@@ -24,7 +26,7 @@ namespace System.Windows.Forms
     [DefaultProperty(nameof(Text))]
     [Serializable]
     [SuppressMessage("Microsoft.Usage", "CA2240:ImplementISerializableCorrectly")]
-    public class ListViewItem : ICloneable, ISerializable
+    public partial class ListViewItem : ICloneable, ISerializable
     {
         private const int MaxSubItems = 4096;
 
@@ -43,7 +45,7 @@ namespace System.Windows.Forms
         private string groupName;
 
         private ListViewSubItemCollection listViewSubItemCollection = null;
-        private ListViewSubItem[] subItems;
+        private ListViewSubItem[] _listViewSubItems;
 
         // we stash the last index we got as a seed to GetDisplayIndex.
         private int lastIndex = -1;
@@ -55,26 +57,6 @@ namespace System.Windows.Forms
         private ListViewItemImageIndexer imageIndexer;
         private string toolTipText = string.Empty;
         private object userData;
-
-        // We need a special way to defer to the ListView's image
-        // list for indexing purposes.
-        internal class ListViewItemImageIndexer : ImageList.Indexer
-        {
-            private readonly ListViewItem _owner;
-
-
-            public ListViewItemImageIndexer(ListViewItem item)
-            {
-                _owner = item;
-            }
-
-
-            public override ImageList ImageList
-            {
-                get => _owner?.ImageList;
-                set => Debug.Fail("We should never set the image list");
-            }
-        }
 
         public ListViewItem()
         {
@@ -111,10 +93,10 @@ namespace System.Windows.Forms
             ImageIndexer.Index = imageIndex;
             if (items != null && items.Length > 0)
             {
-                subItems = new ListViewSubItem[items.Length];
+                _listViewSubItems = new ListViewSubItem[items.Length];
                 for (int i = 0; i < items.Length; i++)
                 {
-                    subItems[i] = new ListViewSubItem(this, items[i]);
+                    _listViewSubItems[i] = new ListViewSubItem(this, items[i]);
                 }
                 SubItemCount = items.Length;
             }
@@ -130,7 +112,7 @@ namespace System.Windows.Forms
         public ListViewItem(ListViewSubItem[] subItems, int imageIndex) : this()
         {
             ImageIndexer.Index = imageIndex;
-            this.subItems = subItems ?? throw new ArgumentNullException(nameof(subItems));
+            this._listViewSubItems = subItems ?? throw new ArgumentNullException(nameof(subItems));
             SubItemCount = subItems.Length;
 
             // Update the owner of these subitems
@@ -192,10 +174,10 @@ namespace System.Windows.Forms
             ImageIndexer.Key = imageKey;
             if (items != null && items.Length > 0)
             {
-                subItems = new ListViewSubItem[items.Length];
+                _listViewSubItems = new ListViewSubItem[items.Length];
                 for (int i = 0; i < items.Length; i++)
                 {
-                    subItems[i] = new ListViewSubItem(this, items[i]);
+                    _listViewSubItems[i] = new ListViewSubItem(this, items[i]);
                 }
                 SubItemCount = items.Length;
             }
@@ -211,7 +193,7 @@ namespace System.Windows.Forms
         public ListViewItem(ListViewSubItem[] subItems, string imageKey) : this()
         {
             ImageIndexer.Key = imageKey;
-            this.subItems = subItems ?? throw new ArgumentNullException(nameof(subItems));
+            this._listViewSubItems = subItems ?? throw new ArgumentNullException(nameof(subItems));
             SubItemCount = subItems.Length;
 
             // Update the owner of these subitems
@@ -268,7 +250,7 @@ namespace System.Windows.Forms
                 }
                 else
                 {
-                    return subItems[0].BackColor;
+                    return _listViewSubItems[0].BackColor;
                 }
             }
             set => SubItems[0].BackColor = value;
@@ -370,7 +352,7 @@ namespace System.Windows.Forms
                 }
                 else
                 {
-                    return subItems[0].Font;
+                    return _listViewSubItems[0].Font;
                 }
             }
             set => SubItems[0].Font = value;
@@ -393,7 +375,7 @@ namespace System.Windows.Forms
                 }
                 else
                 {
-                    return subItems[0].ForeColor;
+                    return _listViewSubItems[0].ForeColor;
                 }
             }
             set
@@ -588,7 +570,7 @@ namespace System.Windows.Forms
                 }
                 else
                 {
-                    return subItems[0].Name;
+                    return _listViewSubItems[0].Name;
                 }
             }
             set => SubItems[0].Name = value;
@@ -751,8 +733,8 @@ namespace System.Windows.Forms
             {
                 if (SubItemCount == 0)
                 {
-                    subItems = new ListViewSubItem[1];
-                    subItems[0] = new ListViewSubItem(this, string.Empty);
+                    _listViewSubItems = new ListViewSubItem[1];
+                    _listViewSubItems[0] = new ListViewSubItem(this, string.Empty);
                     SubItemCount = 1;
                 }
 
@@ -788,7 +770,7 @@ namespace System.Windows.Forms
                 }
                 else
                 {
-                    return subItems[0].Text;
+                    return _listViewSubItems[0].Text;
                 }
             }
             set => SubItems[0].Text = value;
@@ -884,7 +866,7 @@ namespace System.Windows.Forms
             {
                 newItem = (ListViewItem)Activator.CreateInstance(clonedType);
             }
-            newItem.subItems = clonedSubItems;
+            newItem._listViewSubItems = clonedSubItems;
             newItem.ImageIndexer.Index = ImageIndexer.Index;
             newItem.SubItemCount = SubItemCount;
             newItem.Checked = Checked;
@@ -1204,8 +1186,8 @@ namespace System.Windows.Forms
                     newItem.owner = this;
                     newItems[i] = newItem;
                 }
-                newItems[0] = subItems[0];
-                subItems = newItems;
+                newItems[0] = _listViewSubItems[0];
+                _listViewSubItems = newItems;
             }
         }
 
@@ -1225,7 +1207,7 @@ namespace System.Windows.Forms
                 info.AddValue("SubItemCount", SubItemCount);
                 for (int i = 1; i < SubItemCount; i++)
                 {
-                    info.AddValue("SubItem" + i.ToString(CultureInfo.InvariantCulture), subItems[i], typeof(ListViewSubItem));
+                    info.AddValue("SubItem" + i.ToString(CultureInfo.InvariantCulture), _listViewSubItems[i], typeof(ListViewSubItem));
                 }
             }
             info.AddValue("BackColor", BackColor);
@@ -1274,13 +1256,13 @@ namespace System.Windows.Forms
                 int itemIndex = Index;
                 if (index != -1)
                 {
-                    listView.SetItemText(itemIndex, index, subItems[index].Text);
+                    listView.SetItemText(itemIndex, index, _listViewSubItems[index].Text);
                 }
                 else
                 {
                     for (int i = 0; i < subItemCount; i++)
                     {
-                        listView.SetItemText(itemIndex, i, subItems[i].Text);
+                        listView.SetItemText(itemIndex, i, _listViewSubItems[i].Text);
                     }
                 }
 
@@ -1294,702 +1276,6 @@ namespace System.Windows.Forms
         void ISerializable.GetObjectData(SerializationInfo info, StreamingContext context)
         {
             Serialize(info, context);
-        }
-
-        [TypeConverterAttribute(typeof(ListViewSubItemConverter))]
-        [ToolboxItem(false)]
-        [DesignTimeVisible(false)]
-        [DefaultProperty(nameof(Text))]
-        [Serializable]
-        public class ListViewSubItem
-        {
-
-            [NonSerialized]
-            internal ListViewItem owner;
-
-            private string text;
-
-            [OptionalField(VersionAdded = 2)]
-            private string name = null;
-
-            private SubItemStyle style;
-
-            [OptionalField(VersionAdded = 2)]
-            private object userData;
-
-            public ListViewSubItem()
-            {
-            }
-
-            public ListViewSubItem(ListViewItem owner, string text)
-            {
-                this.owner = owner;
-                this.text = text;
-            }
-
-            public ListViewSubItem(ListViewItem owner, string text, Color foreColor, Color backColor, Font font)
-            {
-                this.owner = owner;
-                this.text = text;
-                style = new SubItemStyle
-                {
-                    foreColor = foreColor,
-                    backColor = backColor,
-                    font = font
-                };
-            }
-
-
-            public Color BackColor
-            {
-                get
-                {
-                    if (style != null && style.backColor != Color.Empty)
-                    {
-                        return style.backColor;
-                    }
-
-                    if (owner != null && owner.listView != null)
-                    {
-                        return owner.listView.BackColor;
-                    }
-
-                    return SystemColors.Window;
-                }
-                set
-                {
-                    if (style == null)
-                    {
-                        style = new SubItemStyle();
-                    }
-
-                    if (style.backColor != value)
-                    {
-                        style.backColor = value;
-                        owner?.InvalidateListView();
-                    }
-                }
-            }
-            [Browsable(false)]
-            public Rectangle Bounds
-            {
-                get
-                {
-                    if (owner != null && owner.listView != null && owner.listView.IsHandleCreated)
-                    {
-                        return owner.listView.GetSubItemRect(owner.Index, owner.SubItems.IndexOf(this));
-                    }
-                    else
-                    {
-                        return Rectangle.Empty;
-                    }
-                }
-            }
-
-            internal bool CustomBackColor
-            {
-                get
-                {
-                    Debug.Assert(style != null, "Should have checked CustomStyle");
-                    return !style.backColor.IsEmpty;
-                }
-            }
-
-            internal bool CustomFont
-            {
-                get
-                {
-                    Debug.Assert(style != null, "Should have checked CustomStyle");
-                    return style.font != null;
-                }
-            }
-
-            internal bool CustomForeColor
-            {
-                get
-                {
-                    Debug.Assert(style != null, "Should have checked CustomStyle");
-                    return !style.foreColor.IsEmpty;
-                }
-            }
-
-            internal bool CustomStyle => style != null;
-
-            [Localizable(true)]
-            public Font Font
-            {
-                get
-                {
-                    if (style != null && style.font != null)
-                    {
-                        return style.font;
-                    }
-
-                    if (owner != null && owner.listView != null)
-                    {
-                        return owner.listView.Font;
-                    }
-
-                    return Control.DefaultFont;
-                }
-                set
-                {
-                    if (style == null)
-                    {
-                        style = new SubItemStyle();
-                    }
-
-                    if (style.font != value)
-                    {
-                        style.font = value;
-                        owner?.InvalidateListView();
-                    }
-                }
-            }
-
-            public Color ForeColor
-            {
-                get
-                {
-                    if (style != null && style.foreColor != Color.Empty)
-                    {
-                        return style.foreColor;
-                    }
-
-                    if (owner != null && owner.listView != null)
-                    {
-                        return owner.listView.ForeColor;
-                    }
-
-                    return SystemColors.WindowText;
-                }
-                set
-                {
-                    if (style == null)
-                    {
-                        style = new SubItemStyle();
-                    }
-
-                    if (style.foreColor != value)
-                    {
-                        style.foreColor = value;
-                        owner?.InvalidateListView();
-                    }
-                }
-            }
-
-            [SRCategory(nameof(SR.CatData))]
-            [Localizable(false)]
-            [Bindable(true)]
-            [SRDescription(nameof(SR.ControlTagDescr))]
-            [DefaultValue(null)]
-            [TypeConverter(typeof(StringConverter))]
-            public object Tag
-            {
-                get => userData;
-                set => userData = value;
-            }
-
-            [Localizable(true)]
-            public string Text
-            {
-                get => text ?? string.Empty;
-                set
-                {
-                    text = value;
-                    owner?.UpdateSubItems(-1);
-                }
-            }
-
-            [Localizable(true)]
-            public string Name
-            {
-                get => name ?? string.Empty;
-                set
-                {
-                    name = value;
-                    owner?.UpdateSubItems(-1);
-                }
-            }
-
-            [OnDeserializing]
-            private void OnDeserializing(StreamingContext ctx)
-            {
-            }
-
-            [OnDeserialized]
-            [SuppressMessage("Microsoft.Performance", "CA1801:AvoidUnusedParameters")]
-            private void OnDeserialized(StreamingContext ctx)
-            {
-                name = null;
-                userData = null;
-            }
-
-            [OnSerializing]
-            private void OnSerializing(StreamingContext ctx)
-            {
-            }
-
-            [OnSerialized]
-            private void OnSerialized(StreamingContext ctx)
-            {
-            }
-
-            public void ResetStyle()
-            {
-                if (style != null)
-                {
-                    style = null;
-                    owner?.InvalidateListView();
-                }
-            }
-
-            public override string ToString() => "ListViewSubItem: {" + Text + "}";
-
-            [Serializable]
-            private class SubItemStyle
-            {
-                public Color backColor = Color.Empty;
-                public Color foreColor = Color.Empty;
-                public Font font = null;
-            }
-        }
-
-        public class ListViewSubItemCollection : IList
-        {
-            private readonly ListViewItem _owner;
-
-            // A caching mechanism for key accessor
-            // We use an index here rather than control so that we don't have lifetime
-            // issues by holding on to extra references.
-            private int _lastAccessedIndex = -1;
-
-            public ListViewSubItemCollection(ListViewItem owner)
-            {
-                _owner = owner ?? throw new ArgumentNullException(nameof(owner));
-            }
-
-            /// <summary>
-            /// Returns the total number of items within the list view.
-            /// </summary>
-            [Browsable(false)]
-            public int Count => _owner.SubItemCount;
-
-            object ICollection.SyncRoot => this;
-
-            bool ICollection.IsSynchronized => true;
-
-            bool IList.IsFixedSize => false;
-
-            public bool IsReadOnly => false;
-
-            /// <summary>
-            /// Returns a ListViewSubItem given it's zero based index into the ListViewSubItemCollection.
-            /// </summary>
-            public ListViewSubItem this[int index]
-            {
-                get
-                {
-                    if (index < 0 || index >= Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    return _owner.subItems[index];
-                }
-                set
-                {
-                    if (index < 0 || index >= Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    _owner.subItems[index] = value ?? throw new ArgumentNullException(nameof(value));
-                    _owner.UpdateSubItems(index);
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get => this[index];
-                set
-                {
-                    if (!(value is ListViewSubItem item))
-                    {
-                        throw new ArgumentException(SR.ListViewBadListViewSubItem, nameof(value));
-                    }
-
-                    this[index] = item;
-                }
-            }
-            /// <summary>
-            /// Retrieves the child control with the specified key.
-            /// </summary>
-            public virtual ListViewSubItem this[string key]
-            {
-                get
-                {
-                    // We do not support null and empty string as valid keys.
-                    if (string.IsNullOrEmpty(key))
-                    {
-                        return null;
-                    }
-
-                    // Search for the key in our collection
-                    int index = IndexOfKey(key);
-                    if (!IsValidIndex(index))
-                    {
-                        return null;
-                    }
-
-                    return this[index];
-                }
-            }
-
-            public ListViewSubItem Add(ListViewSubItem item)
-            {
-                if (item == null)
-                {
-                    throw new ArgumentNullException(nameof(item));
-                }
-
-                EnsureSubItemSpace(1, -1);
-                item.owner = _owner;
-                _owner.subItems[_owner.SubItemCount] = item;
-                _owner.UpdateSubItems(_owner.SubItemCount++);
-                return item;
-            }
-
-            public ListViewSubItem Add(string text)
-            {
-                ListViewSubItem item = new ListViewSubItem(_owner, text);
-                Add(item);
-                return item;
-            }
-
-            public ListViewSubItem Add(string text, Color foreColor, Color backColor, Font font)
-            {
-                ListViewSubItem item = new ListViewSubItem(_owner, text, foreColor, backColor, font);
-                Add(item);
-                return item;
-            }
-
-            public void AddRange(ListViewSubItem[] items)
-            {
-                if (items == null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-
-                EnsureSubItemSpace(items.Length, -1);
-                foreach (ListViewSubItem item in items)
-                {
-                    if (item != null)
-                    {
-                        _owner.subItems[_owner.SubItemCount++] = item;
-                    }
-                }
-
-                _owner.UpdateSubItems(-1);
-            }
-
-            public void AddRange(string[] items)
-            {
-                if (items == null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-
-                EnsureSubItemSpace(items.Length, -1);
-                foreach (string item in items)
-                {
-                    if (item != null)
-                    {
-                        _owner.subItems[_owner.SubItemCount++] = new ListViewSubItem(_owner, item);
-                    }
-                }
-
-                _owner.UpdateSubItems(-1);
-            }
-
-            public void AddRange(string[] items, Color foreColor, Color backColor, Font font)
-            {
-                if (items == null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-
-                EnsureSubItemSpace(items.Length, -1);
-                foreach (string item in items)
-                {
-                    if (item != null)
-                    {
-                        _owner.subItems[_owner.SubItemCount++] = new ListViewSubItem(_owner, item, foreColor, backColor, font);
-                    }
-                }
-
-                _owner.UpdateSubItems(-1);
-            }
-
-            int IList.Add(object item)
-            {
-                if (!(item is ListViewSubItem itemValue))
-                {
-                    throw new ArgumentException(SR.ListViewSubItemCollectionInvalidArgument, nameof(item));
-                }
-
-                return IndexOf(Add(itemValue));
-            }
-
-            public void Clear()
-            {
-                int oldCount = _owner.SubItemCount;
-                if (oldCount > 0)
-                {
-                    _owner.SubItemCount = 0;
-                    _owner.UpdateSubItems(-1, oldCount);
-                }
-            }
-
-            public bool Contains(ListViewSubItem subItem) => IndexOf(subItem) != -1;
-
-            bool IList.Contains(object item)
-            {
-                if (!(item is ListViewSubItem itemValue))
-                {
-                    return false;
-                }
-
-                return Contains(itemValue);
-            }
-
-            /// <summary>
-            /// Returns true if the collection contains an item with the specified key, false otherwise.
-            /// </summary>
-            public virtual bool ContainsKey(string key) => IsValidIndex(IndexOfKey(key));
-
-            /// <summary>
-            /// Ensures that the sub item array has the given
-            /// capacity. If it doesn't, it enlarges the
-            /// array until it does. If index is -1, additional
-            /// space is tacked onto the end. If it is a valid
-            /// insertion index into the array, this will move
-            /// the array data to accomodate the space.
-            /// </summary>
-            private void EnsureSubItemSpace(int size, int index)
-            {
-                if (_owner.SubItemCount == ListViewItem.MaxSubItems)
-                {
-                    throw new InvalidOperationException(SR.ErrorCollectionFull);
-                }
-
-                if (_owner.subItems == null || _owner.SubItemCount + size > _owner.subItems.Length)
-                {
-                    // Must grow array. Don't do it just by size, though;
-                    // chunk it for efficiency.
-                    if (_owner.subItems == null)
-                    {
-                        int newSize = (size > 4) ? size : 4;
-                        _owner.subItems = new ListViewSubItem[newSize];
-                    }
-                    else
-                    {
-                        int newSize = _owner.subItems.Length * 2;
-                        while (newSize - _owner.SubItemCount < size)
-                        {
-                            newSize *= 2;
-                        }
-
-                        ListViewSubItem[] newItems = new ListViewSubItem[newSize];
-
-                        // Now, when copying to the member variable, use index
-                        // if it was provided.
-                        if (index != -1)
-                        {
-                            Array.Copy(_owner.subItems, 0, newItems, 0, index);
-                            Array.Copy(_owner.subItems, index, newItems, index + size, _owner.SubItemCount - index);
-                        }
-                        else
-                        {
-                            Array.Copy(_owner.subItems, newItems, _owner.SubItemCount);
-                        }
-                        _owner.subItems = newItems;
-                    }
-                }
-                else
-                {
-                    // We had plenty of room. Just move the items if we need to
-                    if (index != -1)
-                    {
-                        for (int i = _owner.SubItemCount - 1; i >= index; i--)
-                        {
-                            _owner.subItems[i + size] = _owner.subItems[i];
-                        }
-                    }
-                }
-            }
-
-            public int IndexOf(ListViewSubItem subItem)
-            {
-                for (int index = 0; index < Count; ++index)
-                {
-                    if (_owner.subItems[index] == subItem)
-                    {
-                        return index;
-                    }
-                }
-
-                return -1;
-            }
-
-            int IList.IndexOf(object subItem)
-            {
-                if (!(subItem is ListViewSubItem subItemValue))
-                {
-                    return -1;
-                }
-
-                return IndexOf(subItemValue);
-            }
-
-            /// <summary>
-            /// The zero-based index of the first occurrence of value within the entire CollectionBase, if found; otherwise, -1.
-            /// </summary>
-            public virtual int IndexOfKey(string key)
-            {
-                if (string.IsNullOrEmpty(key))
-                {
-                    return -1;
-                }
-
-                if (IsValidIndex(_lastAccessedIndex))
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[_lastAccessedIndex].Name, key, ignoreCase: true))
-                    {
-                        return _lastAccessedIndex;
-                    }
-                }
-
-                for (int i = 0; i < Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[i].Name, key, ignoreCase: true))
-                    {
-                        _lastAccessedIndex = i;
-                        return i;
-                    }
-                }
-
-                _lastAccessedIndex = -1;
-                return -1;
-            }
-
-            /// <summary>
-            /// Determines if the index is valid for the collection.
-            /// </summary>
-            private bool IsValidIndex(int index) => ((index >= 0) && (index < Count));
-
-            public void Insert(int index, ListViewSubItem item)
-            {
-                if (index < 0 || index > Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index));
-                }
-                if (item == null)
-                {
-                    throw new ArgumentNullException(nameof(item));
-                }
-
-                item.owner = _owner;
-
-                EnsureSubItemSpace(1, index);
-
-                // Insert new item
-                _owner.subItems[index] = item;
-                _owner.SubItemCount++;
-                _owner.UpdateSubItems(-1);
-            }
-
-            void IList.Insert(int index, object item)
-            {
-                if (!(item is ListViewSubItem itemValue))
-                {
-                    throw new ArgumentException(SR.ListViewBadListViewSubItem, nameof(item));
-                }
-
-                Insert(index, (ListViewSubItem)item);
-            }
-
-            public void Remove(ListViewSubItem item)
-            {
-                int index = IndexOf(item);
-                if (index != -1)
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            void IList.Remove(object item)
-            {
-                if (item is ListViewSubItem itemValue)
-                {
-                    Remove(itemValue);
-                }
-            }
-
-            public void RemoveAt(int index)
-            {
-                if (index < 0 || index >= Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index));
-                }
-
-                // Remove ourselves as the owner.
-                _owner.subItems[index].owner = null;
-
-                // Collapse the items
-                for (int i = index + 1; i < _owner.SubItemCount; i++)
-                {
-                    _owner.subItems[i - 1] = _owner.subItems[i];
-                }
-
-                int oldCount = _owner.SubItemCount;
-                _owner.SubItemCount--;
-                _owner.subItems[_owner.SubItemCount] = null;
-                _owner.UpdateSubItems(-1, oldCount);
-            }
-
-            /// <summary>
-            /// Removes the child control with the specified key.
-            /// </summary>
-            public virtual void RemoveByKey(string key)
-            {
-                int index = IndexOfKey(key);
-                if (IsValidIndex(index))
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            void ICollection.CopyTo(Array dest, int index)
-            {
-                if (Count > 0)
-                {
-                    Array.Copy(_owner.subItems, 0, dest, index, Count);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                if (_owner.subItems != null)
-                {
-                    return new WindowsFormsUtils.ArraySubsetEnumerator(_owner.subItems, _owner.SubItemCount);
-                }
-                else
-                {
-                    return Array.Empty<ListViewSubItem>().GetEnumerator();
-                }
-
-            }
         }
     }
 }
