+++ src/System.Windows.Forms/src/misc/InvariantComparer.cs
@@ -2,13 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Collections;
+using System.Globalization;
+
 namespace System
 {
-    using System;
-    using System.Collections;
-    using System.Globalization;
-
-    [Serializable]
     internal class InvariantComparer : IComparer
     {
         private readonly CompareInfo m_compareInfo;
@@ -32,4 +30,3 @@ namespace System
         }
     }
 }
-
